+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     - ADD: request body params now is called `body` by default (you can still use `params` and `data`)
     - ADD: returns an error if `body` is not a dictionary
     - ENH: now there is a better error explanation when JSON file is not valid
+    - ENH: now `--version` returns the version number prefixed by `v`
 
 ## v2.1.1